Rename events list
