+++ website/frontend/src/components/pages/Explorer/Events/children/Event/Event.module.scss
