+++ website/frontend/src/components/pages/Explorer/Events/children/Event/children/Body/children/LogContent/LogContent.tsx
@@ -4,7 +4,7 @@ import { Codec } from '@polkadot/types/types';
 import { ProgramModel } from 'types/program';
 import { programService } from 'services/ProgramsRequestService';
 import { getLocalProgram, isDevChain } from 'helpers';
-import { TypeKey } from 'types/events-list';
+import { TypeKey } from 'types/explorer';
 import { Checkbox } from 'common/components/Checkbox/Checkbox';
 import { Pre } from '../Pre/Pre';
 import styles from './LogContent.module.scss';
