+++ website/frontend/src/components/pages/Explorer/Events/children/Event/children/Header/Header.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import clsx from 'clsx';
 import { AnyJson } from '@polkadot/types/types';
 import styles from './Header.module.scss';
-import commonStyles from '../../EventItem.module.scss';
+import commonStyles from '../../Event.module.scss';
 
 type Props = {
   caption: string;
